+++ projects/pastanaga-angular/src/lib/translate/translate.module.ts
@@ -1,7 +1,11 @@
-import { NgModule } from '@angular/core';
+import { ModuleWithProviders, NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
-import { TranslatePipe } from './translate.pipe';
+import { PA_TRANSLATIONS, TranslatePipe } from './translate.pipe';
 import { TranslateDirective } from './translate.directive';
+import { Translation } from './translate.model';
+import { mergeTranslations } from './translate.utils';
+
+const _TRANSLATIONS: Translation = {};
 
 @NgModule({
     imports: [CommonModule],
@@ -9,4 +13,17 @@ import { TranslateDirective } from './translate.directive';
     declarations: [TranslatePipe, TranslateDirective],
     providers: [TranslatePipe],
 })
-export class PaTranslateModule {}
+export class PaTranslateModule {
+    static addTranslations(newTranslations: Translation[]): ModuleWithProviders<PaTranslateModule> {
+        mergeTranslations(_TRANSLATIONS, newTranslations);
+        return {
+            ngModule: PaTranslateModule,
+            providers: [
+                {
+                    provide: PA_TRANSLATIONS,
+                    useFactory: () => _TRANSLATIONS,
+                },
+            ],
+        };
+    }
+}